Change malloc.Lookup to return the size as uintptr rather than uint64.  This changes the Go code to be consistent with the C code.
R=rsc
DELTA=6  (0 added, 0 deleted, 6 changed)
OCL=22983
CL=22987

+++ src/lib/malloc.go
@@ -16,4 +16,4 @@ export type Stats struct {
 export func Alloc(uint64) *byte;
 export func Free(*byte);
 export func GetStats() *Stats;
-export func Lookup(*byte) (*byte, uint64);
+export func Lookup(*byte) (*byte, uintptr);
